+++ pages/ai-search.tsx
@@ -44,7 +44,7 @@ const AiSearch: NextPage = () => {
     queryKey: ["search", useDebouncedValue(search, 200)] as const, // Menyesuaikan tipe argumen queryKey
     queryFn: () => searchData(search),
     enabled: enabled,
-    cacheTime: 1000 * 60 * 60 * 24,
+    gcTime: 1000 * 60 * 60 * 24,
   });
 
   return (